moves comp function to own mod
