+++ mal_rustc_core/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod parser;
 pub mod types;
+pub mod comp;
 
 pub static MAL_RUNTIME: &str = include_str!("types/runtime.rs");
